requirements.txt
kivy==2.1.0
kivymd==1.1.1
supabase
requests
pyjnius